Add files via upload
This C program prints the multiplication table for a specified number up to a specified number of times. The user is prompted to enter the number for which the multiplication table is to be generated and the number of times the table should be printed.

Features
Accepts a number from the user to generate its multiplication table.
Accepts the number of times the table should be printed.
Displays the multiplication table in a formatted manner.
How It Works
The program starts by displaying a title message.
It prompts the user to enter the number for which the multiplication table is to be generated.
It then prompts the user to enter how many times the table should be printed.
An array is created to store the results of the multiplication.
A loop calculates the multiplication results and stores them in the array.
Another loop prints the multiplication table in a formatted way.